chore: rename utility functions file
